First attempt at fair queuing with PIFOs.
Surprisingly difficult because PIFOs don't support state updates on dequeue.
Seems like a design flaw.

Still has debugging info enabled. Needs to be removed once things have been
better sanity checked.

+++ code/bess/core/traffic_class.h
@@ -74,7 +74,7 @@ enum resource_t {
 typedef uint64_t resource_arr_t[NUM_RESOURCES];
 
 // The priority of a traffic class.
-typedef uint32_t priority_t;
+typedef uint64_t priority_t;
 
 // The amount of a resource allocated to a class.
 typedef int32_t resource_share_t;
